refactor: migrating jest module mocks to typescript
